requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google APIs and AI
google-api-python-client>=2.110.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.1.1
google-generativeai>=0.8.3
google-genai>=0.3.0

# Background tasks and scheduling
apscheduler>=3.10.4

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Image processing
Pillow>=10.1.0

# Cloud storage (Cloudflare R2 / S3 compatible)
boto3>=1.34.0

# Database
psycopg2-binary>=2.9.9

# Additional utilities used in codebase
typing-extensions>=4.8.0
python-multipart>=0.0.6
